# gin-petclinic-service/docs/swagger.yaml
basePath: /v1
definitions:
  errors.ErrorResponse:
    properties:
      details: {}
      message:
        example: status bad request
        type: string
      status:
        example: 400
        type: integer
    type: object
  model.Context:
    properties:
      count:
        type: integer
    type: object
...
  pet.Response:
    properties:
      birthdate:
        type: string
      id:
        type: integer
      name:
        type: string
      type:
        type: string
      visits:
        items:
          $ref: '#/definitions/visit.Response'
        type: array
    type: object
  pet.Responses:
    properties:
      context:
        $ref: '#/definitions/model.Context'
      pets:
        items:
          $ref: '#/definitions/pet.Response'
        type: array
    type: object
...
  vet.Response:
    properties:
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      specialties:
        items:
          $ref: '#/definitions/vet.specialtyResponse'
        type: array
    type: object
  vet.Responses:
    properties:
      context:
        $ref: '#/definitions/model.Context'
      vets:
        items:
          $ref: '#/definitions/vet.Response'
        type: array
    type: object
  vet.specialtyResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  visit.Response:
    properties:
      description:
        type: string
      id:
        type: integer
      visitDate:
        type: string
    type: object
  visit.Responses:
    properties:
      context:
        $ref: '#/definitions/model.Context'
      visits:
        items:
          $ref: '#/definitions/visit.Response'
        type: array
    type: object
  visit.UpdateRequest:
    properties:
      description:
        type: string
      id:
        type: integer
      petId:
        type: integer
      visitDate:
        type: string
    required:
    - description
    - petId
    - visitDate
    type: object
...
host: localhost:8443
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a pet clinic API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pet Clinic API
  version: "1.0"
paths:
...
  /pets/{id}/visits:
    get:
      description: Get pet with visits by ID
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pet.Responses'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get pet with visits by ID
      tags:
      - pets
...
  /vets/all:
    get:
      description: Get all vets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vet.Responses'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: List all vets
      tags:
      - vets
  /visits/{id}:
    get:
      description: Get visit by ID
      parameters:
      - description: Visit ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/visit.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get visit by id
      tags:
      - visits
    put:
      description: Update visit
      parameters:
      - description: Visit ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update visit
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/visit.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/visit.Responses'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Update visit
      tags:
      - visits
  /visits/all:
    get:
      description: Get all visits
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/visit.Responses'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: List all visits
      tags:
      - visits
schemes:
- https
swagger: "2.0"